+++ tests/unit/ConfigTest.php
@@ -6,26 +6,24 @@ namespace davidhirtz\yii2\config\tests\unit;
 
 use Codeception\Test\Unit;
 use davidhirtz\yii2\config\modules\admin\models\Config;
-use davidhirtz\yii2\config\modules\admin\Module;
+use davidhirtz\yii2\config\tests\support\UnitTester;
 use Yii;
 use yii\db\AfterSaveEvent;
 use yii\db\BaseActiveRecord;
-use yii\helpers\FileHelper;
 
 class ConfigTest extends Unit
 {
-    private ?string $_configFile = null;
-    private bool $isEventTriggered = false;
+    protected UnitTester $tester;
 
     protected function _before(): void
     {
-        $this->deleteConfigFile();
+        $this->tester->deleteConfigFile();
         parent::_before();
     }
 
     protected function _after(): void
     {
-        $this->deleteConfigFile();
+        $this->tester->deleteConfigFile();
         parent::_after();
     }
 
@@ -40,44 +38,73 @@ class ConfigTest extends Unit
 
         self::assertTrue($config->save());
         self::assertEquals('unit-test', Yii::$app->params['cookieValidationKey']);
-        self::assertFileExists(Yii::getAlias($this->getConfigFile()));
+        self::assertFileExists(Yii::getAlias($this->tester->getConfigFile()));
 
         $config = TestConfig::create();
         self::assertEquals('unit-test', $config->cookieValidationKey);
     }
 
-    public function testUpdateConfigWithEvent()
+    public function testUpdateConfig()
     {
         $config = TestConfig::create();
         $config->cookieValidationKey = 'unit-test';
-        $config->save();
+
+        self::assertTrue($config->save());
 
         $config->cookieValidationKey = 'unit-test-2';
 
-        $config->on(BaseActiveRecord::EVENT_AFTER_UPDATE, function (AfterSaveEvent $event) {
+        $isTriggered = false;
+
+        $config->on(BaseActiveRecord::EVENT_AFTER_UPDATE, function (AfterSaveEvent $event) use (&$isTriggered) {
             self::assertEquals(['cookieValidationKey' => 'unit-test'], $event->changedAttributes);
-            $this->isEventTriggered = true;
+            $isTriggered = true;
         });
 
         self::assertTrue($config->save());
-        self::assertTrue($this->isEventTriggered);
+        self::assertTrue($isTriggered);
+
+        self::assertEquals('unit-test-2', $config->cookieValidationKey);
+        self::assertEquals('unit-test-2', Yii::$app->params['cookieValidationKey']);
     }
 
-    protected function getConfigFile(): string
+    public function testSaveI18nAttribute()
     {
-        if ($this->_configFile === null) {
-            /** @var Module $module */
-            $module = Yii::$app->getModule('admin')->getModule('config');
-            $this->_configFile = $module->configFile;
-        }
+        Yii::$app->getI18n()->languages = ['de', 'en-US'];
 
-        return $this->_configFile;
+        $config = TestConfigI18n::create();
+        $config->cookieValidationKey = 'unit-test';
+
+        self::assertFalse($config->save());
+        self::assertArrayHasKey('cookieValidationKey_de', $config->getErrors());
+
+        $config->cookieValidationKey_de = 'unit-test-de';
+
+        self::assertTrue($config->save());
+
+        self::assertEquals('unit-test-de', $config->cookieValidationKey_de);
+        self::assertEquals('unit-test-de', Yii::$app->params['cookieValidationKey_de']);
+
     }
 
-    protected function deleteConfigFile(): void
+    public function testTrailIntegration(): void
     {
-        $file = Yii::getAlias($this->getConfigFile());
-        FileHelper::removeDirectory(dirname($file));
+        $config = TestConfig::create();
+        $config->cookieValidationKey = 'trail-test';
+        $config->save();
+
+        $trail = $this->tester->loadLastTrail();
+
+        self::assertEquals(TestConfig::class, $trail->model);
+        self::assertEquals($config->getTrailModelName(), $trail->getModelName());
+        self::assertEquals(['cookieValidationKey' => [null, 'trail-test']], $trail->data);
+
+        $config->cookieValidationKey = 'trail-test-2';
+        $config->save();
+
+        $trail = $this->tester->loadLastTrail();
+
+        self::assertEquals(TestConfig::class, $trail->model);
+        self::assertEquals(['cookieValidationKey' => ['trail-test', 'trail-test-2']], $trail->data);
     }
 }
 
@@ -88,11 +115,19 @@ class TestConfig extends Config
 {
     public function rules(): array
     {
-        return [
+        return $this->getI18nRules([
             [
                 ['cookieValidationKey'],
                 'required',
             ],
-        ];
+        ]);
     }
+}
+
+/**
+ * @property string $cookieValidationKey_de
+ */
+class TestConfigI18n extends TestConfig
+{
+    public array $i18nAttributes = ['cookieValidationKey'];
 }